superspreading: Understand Individual-Level Variation in Infectious Disease Transmission

Estimate and understand individual-level variation in transmission. Implements density and cumulative compound Poisson discrete distribution functions (Kremer et al. (2021) doi:10.1038/s41598-021-93578-x ), as well as functions to calculate infectious disease outbreak statistics given epidemiological parameters on individual-level transmission; including the probability of an outbreak becoming an epidemic/extinct (Kucharski et al. (2020) doi:10.1016/S1473-3099(20)30144-4 ), or the cluster size statistics, e.g. what proportion of cases cause X% of transmission (Lloyd-Smith et al. (2005) doi:10.1038/nature04153 ).

See also

Author

Maintainer: Joshua W. Lambert joshua.lambert@lshtm.ac.uk (ORCID) [copyright holder]

Authors:

Adam Kucharski adam.kucharski@lshtm.ac.uk (ORCID) [copyright holder]

Dillon C. Adam dcadam@hku.hk (ORCID)

Other contributors:

Sebastian Funk sebastian.funk@lshtm.ac.uk (ORCID) (.chain_sim uses code from bpmodels::chain_sim) [contributor, copyright holder]

Pratik Gupte pratik.gupte@lshtm.ac.uk (ORCID) [reviewer]

Hugo Gruson hugo@data.org (ORCID) [reviewer]

James M. Azam james.azam@lshtm.ac.uk (ORCID) [reviewer, contributor]

Chris Hartgerink chris@data.org (ORCID) [reviewer]