Estimate what proportion of cases that cause a certain proportion of transmission

Calculates the proportion of cases that cause a certain percentage of transmission.

It is commonly estimated what proportion of cases cause 80% of transmission (i.e. secondary cases). This can be calculated using proportion_transmission() at varying values of \(R\) and for different values of percentage transmission.

Usage

proportion_transmission(
  R,
  k,
  percent_transmission,
  sim = FALSE,
  ...,
  offspring_dist
)

Arguments

R

A number specifying the R parameter (i.e. average secondary cases per infectious individual).

k

A number specifying the k parameter (i.e. overdispersion in offspring distribution from fitted negative binomial).

percent_transmission

A number of the percentage transmission for which a proportion of cases has produced.

sim

A logical whether the calculation should be done numerically (i.e. simulate secondary contacts) or analytically.

...

dots not used, extra arguments supplied will cause a warning.

offspring_dist

An <epidist> object. An S3 class for working with epidemiological parameters/distributions, see epiparameter::epidist().

Value

A <data.frame> with the value for the proportion of cases for a given value of R and k.

Details

Multiple values of R and k can be supplied and a <data.frame> of every combination of these will be returned.

The numerical calculation uses random number generation to simulate secondary contacts so the answers may minimally vary between calls. The number of simulation replicates is fixed to 1e5.

References

The analytical calculation is from:

Endo, A., Abbott, S., Kucharski, A. J., & Funk, S. (2020) Estimating the overdispersion in COVID-19 transmission using outbreak sizes outside China. Wellcome Open Research, 5. doi:10.12688/wellcomeopenres.15842.3

Examples

# example of single values of R and k
percent_transmission <- 0.8 # 80% of transmission
R <- 2
k <- 0.5
proportion_transmission(
  R = R,
  k = k,
  percent_transmission = percent_transmission
)
#>   R   k prop_80
#> 1 2 0.5   26.4%

# example with multiple values of k
k <- c(0.1, 0.2, 0.3, 0.4, 0.5, 1)
proportion_transmission(
  R = R,
  k = k,
  percent_transmission = percent_transmission
)
#>   R   k prop_80
#> 1 2 0.1    9.2%
#> 2 2 0.2   15.6%
#> 3 2 0.3   20.3%
#> 4 2 0.4   23.8%
#> 5 2 0.5   26.4%
#> 6 2 1.0   35.6%

# example with vectors of R and k
R <- c(1, 2, 3)
proportion_transmission(
  R = R,
  k = k,
  percent_transmission = percent_transmission
)
#>    R   k prop_80
#> 1  1 0.1    8.7%
#> 2  2 0.1    9.2%
#> 3  3 0.1    9.4%
#> 4  1 0.2   14.3%
#> 5  2 0.2   15.6%
#> 6  3 0.2     16%
#> 7  1 0.3   18.2%
#> 8  2 0.3   20.3%
#> 9  3 0.3     21%
#> 10 1 0.4   20.8%
#> 11 2 0.4   23.8%
#> 12 3 0.4     25%
#> 13 1 0.5   22.6%
#> 14 2 0.5   26.4%
#> 15 3 0.5     28%
#> 16 1 1.0     30%
#> 17 2 1.0   35.6%
#> 18 3 1.0   37.8%